fix(apply_patch,ui,v0): Display "Apply patch" properly for v0 conversat.
v0 conversations automatically look up the localized text by action type
and don't provide an existing switch/case statement injection point to
dynamically provide arguments to the localized text. At the same time,
it's really not necessary for us to make the title dynamic because a
patch can theoretically touch multiple files, so it's best to expose
this info in the action details.